feat: add recipe, film, weather api impl

+++ app/domain/src/main/java/com/fishfromsandiego/whattodo/domain/chore/usecase/EditExistingChore.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class EditExistingChore @Inject constructor(
     val repository: ChoreRepository
 ) {
-    suspend operator fun invoke(id: Int, newValue: ChoreModel) =
+    suspend operator fun invoke(id: Long, newValue: ChoreModel) =
         repository.updateChore(id, newValue)
 }
